farmdata_backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.8.4
aiohttp-retry==2.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==24.4.2
Automat==24.8.1
autopep8==1.6.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.3
bs4==0.0.2
build==1.2.2.post1
CacheControl==0.14.2
certifi==2023.5.7
cffi==1.17.0
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.1.0
cleo==2.1.0
colorama==0.4.6
constantly==23.10.4
crashtest==0.4.1
cryptography==43.0.0
daphne==4.1.2
defusedxml==0.8.0rc2
distlib==0.3.9
distro==1.9.0
Django==4.0.6
django-cors-headers==3.13.0
django-filter==24.3
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djoser==2.2.3
docx2txt==0.8
dulwich==0.22.7
exceptiongroup==1.2.0
fastjsonschema==2.21.1
filelock==3.17.0
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
hyperlink==21.0.0
idna==3.4
importlib_metadata==8.6.1
incremental==24.7.2
infobip-api-python-client==5.0.0
infobip-api-python-sdk==5.0.1
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jiter==0.7.1
keyring==25.6.0
more-itertools==10.6.0
msgpack==1.1.0
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.7
packaging==24.2
paynow==1.0.8
Pillow==9.2.0
pkginfo==1.12.0
platformdirs==4.3.6
poetry==2.0.1
poetry-core==2.0.1
propcache==0.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.8.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.4.0
pyOpenSSL==24.2.1
PyPDF2==3.0.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2022.1
pywin32-ctypes==0.2.3
RapidFuzz==3.12.1
regex==2022.3.15
reportlab==4.3.0
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
service-identity==24.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.2
tomlkit==0.13.2
tqdm==4.65.0
trove-classifiers==2025.1.15.22
twilio==9.4.4
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2022.1
urllib3==2.0.2
virtualenv==20.29.2
yarl==1.9.2
zipp==3.21.0
zope.interface==7.1.1
